Adds code for OilDrums to check explosions

+++ headers/Game.h
@@ -133,7 +133,7 @@ short Game_retreatTimer=0;
 short Game_graceTimer=0;
 
 // sudden death timer - 10 minutes before sudden death begins
-short Game_suddenDeathTimer=((60*100) * 10);
+long Game_suddenDeathTimer=(long)((long)((long)60*(long)100) * (long)10);
 
 // the current team, and the current selected worm on each team:
 char Game_currentTeam=1;
